feature(delete-user): endpoint for admin to delete users admin account can not be deleted write unit and feature test

+++ app/Http/Services/UserService.php
@@ -43,6 +43,10 @@ class UserService {
             return ['status' => false];
         }
 
+        if ($user->is_admin) {
+            return ['status' => false];
+        }
+
         $user->delete();
 
         return ['status' => true];
